+++ src/lib/plugin/sdk/cm_plugin_run_task.rs
@@ -9,7 +9,7 @@ mod cm_plugin_run_task_test;
 
 use crate::runner;
 use crate::types::{FlowInfo, FlowState, RunTaskOptions, Step};
-use duckscript::types::command::{Command, CommandArgs, CommandResult};
+use duckscript::types::command::{Command, CommandInvocationContext, CommandResult};
 use std::cell::RefCell;
 use std::rc::Rc;
 
@@ -29,7 +29,7 @@ impl Command for CommandImpl {
         Box::new((*self).clone())
     }
 
-    fn run(&self, _arguments: CommandArgs) -> CommandResult {
+    fn run(&self, _context: CommandInvocationContext) -> CommandResult {
         let options = RunTaskOptions {
             plugins_enabled: false,
         };